Implement a new transaction format (v4): FlexTrans
This implements the parsing and serialization of v4 transactions
in the format called Compact Message Format.

This fixes the malleability issues in transactions.
This actually shrinks most transactions a couple of procent.
We get linear scaling of sighash operations

Future;
This enables the lightning network to develop.

Next it makes it very easy to fix various problems we know of in Bitcoin;
* far simpler script upgrades.
* fraud proofs.
* pruning signature data in blocks to remove about 75% of the block data.
* reduces bandwidth for light nodes and historical sync

Still work in progress is CSV support.