Update code to reflect the new protocol-upgrade
This reverts BIP109. Classic will no longer vote for BIP109.

This reuses the code and starts voting for a new bit which is a protocol
upgrade and the only change in that one is that the maximum block size is
removed.

This new fork is activated using bit 26, while requiring signalling for
BIP9 making the block version 0x28000000 When 85% of 1000 blocks have this
bit set we lock in this change and have a 4 week wait period.
After that period blocks no longer have a size limit.

Miners can continue to use the -blockmaxsize argument or bitcoin.conf
setting to limit the size of blocks they generate!  All nodes can also use
the `-blocksizeacceptlimit` option to limit the size of the blocks accepted
from other miners.