Introduce new config option blocksizeacceptlimit
This option allows a user to configure the behaviour of the
block-acceptance algorithm. A node will reject incoming blocks that are
larger than the block size accept limit and as such a miner would be able
to decide to limit blocks it will build on top of based on size.

The blocksizeacceptlimit is a number in 100KB increments.
The default is 20, which makes the default accepted block 2MB.